NestedLoops
Wrote the pseudocode which I just suck at and wrote out basic code. Are we like allowed to do more to it? I would like to use the Scanner class to make it so you can choose the size and then would like to add a do while loop so we can choose to do it again.